+++ mirtop/gff/header.py
@@ -1,16 +1,16 @@
 import mirtop.libs.logger as mylog
 logger = mylog.getLogger(__name__)
 
-def create(bam_file, samples, database, filter = None):
+def create(samples, database, custom, filter = None):
     header = ""
     header += _get_gff_version()
     header += _get_database(database)
     header += _get_samples(samples)
+    header += custom
     return header
 
-
 def _get_gff_version():
-    return "# GFF3 adapted for miRNA sequencing data\n## VERSION 0.0.1\n"
+    return "## GFF3 adapted for miRNA sequencing data. VERSION 0.0.1\n"
 
 def _get_samples(samples):
     return "## COLDATA: %s" % ";".join(samples)