+++ mirtop/libs/parse.py
@@ -42,6 +42,8 @@ def add_subparser_gff(subparsers):
                         help="species")
     parser.add_argument("--hairpin", help="hairpin.fa")
     parser.add_argument("--gtf", help="gtf file with precursor and mature position to genome.")
+    parser.add_argument("--format", help="Input format, default BAM file.",
+                        choices=['BAM', 'seqbuster'], default="BAM")
     parser = _add_debug_option(parser)
     return parser